feat: Add additional module utilities for quick wins
- Extract route availability checking to reusable utility
- Add dynamic tab processing helper function
- Simplify ProjectTabs.vue dynamic route logic (85% reduction)
- Standardize module route middleware patterns
- Move DynamicTabOptions type to shared utility

Quick wins achieved:
- 21 lines → 3 lines in ProjectTabs dynamic processing (85% reduction)
- Reusable checkModuleAvailability function
- Consistent route middleware patterns across modules
