+++ packages/frontend/editor-ui/src/features/dataStore/dataStore.routes.ts
@@ -4,17 +4,10 @@ import {
 	DATA_STORE_VIEW,
 	PROJECT_DATA_STORES,
 } from '@/features/dataStore/constants';
-import { useSettingsStore } from '@/stores/settings.store';
+import { checkModuleAvailability, createModuleMiddleware } from '@/utils/module/routeUtils';
 
 const MainSidebar = async () => await import('@/components/MainSidebar.vue');
 const DataStoreView = async () => await import('@/features/dataStore/DataStoreView.vue');
-
-export const checkModuleAvailability = (to?: RouteLocationNormalized): boolean => {
-	if (!to?.meta?.moduleName || typeof to.meta.moduleName !== 'string') {
-		return true;
-	}
-	return useSettingsStore().isModuleActive(to.meta.moduleName);
-};
 
 export const dataStoreRoutes: RouteRecordRaw[] = [
 	{
